implement mail api and refactored admin page
