+++ MonthlyIB/front/components/adminComponents/AdminUser.jsx
@@ -6,16 +6,14 @@ import {
   faEnvelope,
 } from "@fortawesome/free-solid-svg-icons";
 import { useUserInfo, useUserStore } from "@/store/user";
-import { useEffect, useState, useRef } from "react";
+import { useEffect, useState } from "react";
 import AdminUserDetail from "./AdminUserDetail";
+import AdminUserAdminModal from "./AdminUserAdminModal";
+import AdminScheduleModal from "./AdminSchedulemodal";
 import shortid from "shortid";
 import { useSubscribeStore } from "@/store/subscribe";
-import {
-  subscribeGetUserInfo,
-  subscribePostUser,
-  subscribeReviseUser,
-} from "@/apis/subscribeAPI";
 import { getKnitSubscribeDataList } from "@/utils/utils";
+import { mailPost } from "@/apis/mail";
 // 15명씩
 // Refactoring 할것.
 const AdminUser = () => {
@@ -25,19 +23,10 @@ const AdminUser = () => {
   const { subscribeList } = useSubscribeStore();
   const [modal, setModal] = useState(false);
   const [adminModal, setAdminModal] = useState(false);
+  const [mailModal, setMailModal] = useState(false);
+  const [detail, setDetail] = useState("");
   const [authority, setAuthority] = useState("");
   const [subscirbeDataList, setSubscribeDataList] = useState({});
-  const [subscribeId, setSubscribeId] = useState(-1);
-  const [subscribeTitle, setSubscribeTitle] = useState("");
-  const [questionCount, setQuestionCount] = useState("");
-  const [tutoringCount, setTutoringCount] = useState("");
-  const [subscribeMonthPeriod, setSubscribeMonthPeriod] = useState("");
-  const [videoLessonsCount, setVideoLessonsCount] = useState("");
-  const closeRef = useRef();
-
-  const onChangeMonth = (e) => {
-    setSubscribeId(e.target.value);
-  };
 
   useEffect(() => {
     getKnitSubscribeDataList(subscribeList, setSubscribeDataList);
@@ -54,65 +43,14 @@ const AdminUser = () => {
     getUserInfo(userId, userInfo);
   };
 
-  const onSubmitChangeAuthority = async () => {
-    if (authority !== userDetailInfo?.authority) {
-      reviseUserInfo(
-        userDetailInfo?.userId,
-        userDetailInfo?.email,
-        userDetailInfo?.nickName,
-        userDetailInfo?.birth,
-        userDetailInfo?.school,
-        userDetailInfo?.grade,
-        userDetailInfo?.address,
-        userDetailInfo?.country,
-        userDetailInfo?.userStatus,
-        authority,
-        userDetailInfo?.memo,
-        userDetailInfo?.marketingTermsCheck,
-        userInfo
-      );
-      getUserInfo(userDetailInfo?.userId, userInfo);
-    }
+  const onSubmitMail = () => {
+    setMailModal(false);
+    mailPost(userDetailInfo?.userId, detail, userInfo);
+  };
 
-    if (subscribeId !== -1) {
-      const res = await subscribePostUser(
-        userDetailInfo?.userId,
-        subscribeId,
-        userInfo
-      );
-      subscribeReviseUser(
-        res?.data.subscribeUserId,
-        questionCount,
-        tutoringCount,
-        subscribeMonthPeriod,
-        videoLessonsCount,
-        [],
-        userInfo
-      );
-    } else {
-      if (
-        questionCount !== "" ||
-        tutoringCount !== "" ||
-        subscribeMonthPeriod !== "" ||
-        videoLessonsCount !== ""
-      ) {
-        const res = await subscribeGetUserInfo(
-          userDetailInfo?.userId,
-          0,
-          userInfo
-        );
-        subscribeReviseUser(
-          res?.data[0]?.subscribeUserId,
-          questionCount,
-          tutoringCount,
-          subscribeMonthPeriod,
-          videoLessonsCount,
-          [],
-          userInfo
-        );
-      }
-    }
-    setAdminModal(false);
+  const onClickMail = (userId) => {
+    setMailModal(!mailModal);
+    getUserInfo(userId, userInfo);
   };
 
   return (
@@ -148,7 +86,10 @@ const AdminUser = () => {
                       icon={faUserGear}
                       onClick={() => onClickChangeAuthority(v.userId)}
                     />
-                    <FontAwesomeIcon icon={faEnvelope} />
+                    <FontAwesomeIcon
+                      icon={faEnvelope}
+                      onClick={() => onClickMail(v.userId)}
+                    />
                   </span>
                 </div>
                 <hr />
@@ -163,112 +104,31 @@ const AdminUser = () => {
             setModal={setModal}
           />
         )}
-        {adminModal === true && (
-          <div className={styles.md}>
-            <div
-              className={styles.md_box_flex}
-              ref={closeRef}
-              onClick={(e) =>
-                closeRef.current === e.target && setAdminModal(false)
-              }
-            >
-              <div className={styles.admin_box}>
-                <div className={styles.md_top}>
-                  <div className={styles.tit}>
-                    {userDetailInfo?.username} 님의 권한 변경 및 구독 관리
-                  </div>
-                  <select
-                    value={authority}
-                    onChange={(e) => setAuthority(e.target.value)}
-                    disabled={userDetailInfo?.userId === 1 ? true : false}
-                  >
-                    <option value="USER">일반유저</option>
-                    <option value="ADMIN">관리자</option>
-                  </select>
-                  <div>
-                    <span>구독 상품 </span>
-                    <select
-                      value={subscribeTitle}
-                      onChange={(e) => setSubscribeTitle(e.target.value)}
-                    >
-                      <option value="">구독 상품</option>
-                      {Object.keys(subscirbeDataList).map((v, i) => (
-                        <option value={v} key={i}>
-                          {v}
-                        </option>
-                      ))}
-                    </select>
 
-                    <select value={subscribeId} onChange={onChangeMonth}>
-                      <option value={-1}>-</option>
-                      {subscirbeDataList[subscribeTitle]?.map((v, i) => (
-                        <option value={v.subscriberId} key={i}>
-                          {v.subscribeMonthPeriod} 개월
-                        </option>
-                      ))}
-                    </select>
-                  </div>
-                  <div className={styles.subscribe_price_flex}>
-                    <span>질문 수</span>
-                    <input
-                      type="number"
-                      autoFocus="true"
-                      autoComplete="off"
-                      required="Y"
-                      placeholder="질문 갯수"
-                      value={questionCount}
-                      onChange={(e) => setQuestionCount(e.target.value)}
-                    />
-                  </div>
-                  <div className={styles.subscribe_price_flex}>
-                    <span>튜터링 갯수</span>
-                    <input
-                      type="number"
-                      autoFocus="true"
-                      autoComplete="off"
-                      required="Y"
-                      placeholder="튜터링 갯수"
-                      value={tutoringCount}
-                      onChange={(e) => setTutoringCount(e.target.value)}
-                    />
-                  </div>
-                  <div className={styles.subscribe_price_flex}>
-                    <span>구독개월 수</span>
-                    <input
-                      type="number"
-                      autoFocus="true"
-                      autoComplete="off"
-                      required="Y"
-                      placeholder="구독개월"
-                      value={subscribeMonthPeriod}
-                      onChange={(e) => setSubscribeMonthPeriod(e.target.value)}
-                    />
-                  </div>
-                  <div className={styles.subscribe_price_flex}>
-                    <span>영상강의 수</span>
-                    <input
-                      type="number"
-                      autoFocus="true"
-                      autoComplete="off"
-                      required="Y"
-                      placeholder="영상강의 수"
-                      value={videoLessonsCount}
-                      onChange={(e) => setVideoLessonsCount(e.target.value)}
-                    />
-                  </div>
-                </div>
-                <button
-                  type="button"
-                  className={styles.md_btn}
-                  onClick={onSubmitChangeAuthority}
-                >
-                  확인
-                </button>
-              </div>
-            </div>
-            <div className={styles.md_dim}></div>
-          </div>
-        )}
+        <AdminUserAdminModal
+          adminModal={adminModal}
+          setAdminModal={setAdminModal}
+          authority={authority}
+          setAuthority={setAuthority}
+          subscirbeDataList={subscirbeDataList}
+          userDetailInfo={userDetailInfo}
+          userInfo={userInfo}
+          getUserInfo={getUserInfo}
+          reviseUserInfo={reviseUserInfo}
+        />
+
+        <AdminScheduleModal
+          modal={mailModal}
+          setModal={setMailModal}
+          status={null}
+          title={"메일 보내기"}
+          requestUsername={userDetailInfo?.username}
+          detail={detail}
+          setDetail={setDetail}
+          setStatus={null}
+          onSubmitChange={onSubmitMail}
+          onSubmitDelete={null}
+        />
       </div>
     </>
   );
